Restructured routing and pages
