+++ react/.gitignore
@@ -14,6 +14,8 @@ package-lock.json
 
 # misc
 .DS_Store
+.vscode
+.idea
 .env.local
 .env.development.local
 .env.test.local
