+++ react/src/components/Footer/footer.css
@@ -1,4 +1,4 @@
-.love{
+.love {
     display: inline;
     color: red;
 }
