+++ react/src/components/Footer/index.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import './footer.css';
+
 const Footer = () => (
   <>
     <footer className="footer mt-3">
