+++ react/src/components/index.js
@@ -1,0 +1,6 @@
+export {default as SideBar} from './SideBar';
+export {default as Footer} from './Footer';
+export {default as ProfileForm} from './ProfileForm';
+export {default as SignupForm} from './SignupForm';
+export {default as NavBar} from './NavBar';
+export {default as LoginForm} from './LoginForm';
