+++ react/src/index.js
@@ -2,12 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import './home.css';
-import App from './App';
+import Router from './router';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Router/>
   </React.StrictMode>,
   document.getElementById('root')
 );
