+++ react/src/pages/Auth/Login/index.js
@@ -1,35 +1,40 @@
 import React from 'react';
-import '../style/auth.css';
-import SignupForm from "../components/SignupForm";
+import LoginForm from "../../../components/LoginForm";
+import '../auth.css';
 import {Link} from "react-router-dom";
 
-const SignupPage = () => (
-    <div className="container-flex h-100 min-vh-100">
+const LoginPage = () => (
+    <div className="container-flex h-100">
         <div className="row justify-content-center h-100">
             <div className="col-lg-12 col-md-9 h-100">
                 <div className="o-hidden border-0 shadow-lg h-100">
                     <div className="card-body p-0 h-100">
                         <div className="row d-flex h-100">
-                            <div className="col-lg-5 d-none d-lg-block bg-image-register">
+                            <div className="col-lg-5 d-none d-lg-block bg-image-login">
                             </div>
-                            <div className="col-lg-7  col-md-12 p-5 h-100">
+                            <div className="col-lg-7 p-5 h-100">
                                 <div className="p-5">
                                     <div className="text-center mb-5">
                                         <div className="title">PEER RATER</div>
                                     </div>
-                                    <div className="mb-1 just">
-                                        <div className="sign-in-text">Sign Up</div>
+                                    <div className="mb-1">
+                                        <div className="sign-in-text">Sign in</div>
                                     </div>
                                     <div className="mb-5">
                                         <div className="sign-in-description">Enter your email and
-                                            password to sign up
+                                            password to sign In
                                         </div>
                                     </div>
-                                    <SignupForm/>
+                                    <LoginForm/>
                                     <hr/>
-                                    <div className="text-center mt-5">
-                                        <p className="text-muted dont-have-account-message">Already have an account? <a
-                                                className="sign-up-link"><Link to="/login">Sign in</Link></a></p>
+                                    <div className="text-center mb-2">
+
+                                        <a className="sign-up-link"><Link to="/forgotpassword">Forgot Password?</Link></a>
+                                    </div>
+                                    <div className="text-center">
+                                        <p className="text-muted dont-have-account-message">Don't have an
+                                            account? <a
+                                                className="sign-up-link"><Link to="/signup">Sign up</Link></a></p>
                                     </div>
                                 </div>
                             </div>
@@ -40,4 +45,5 @@ const SignupPage = () => (
         </div>
     </div>
 );
-export default SignupPage;
+
+export default LoginPage;
