+++ react/src/pages/Auth/auth.css
@@ -32,17 +32,17 @@ body {
 
 }
 .bg-image-login {
-    background: url("srcackground1.png") center;
+    /* background: url("srcackground1.png") center; */
     background-size: cover;
 }
 
 .bg-image-register {
-    background: url("srcackground2.png") center;
+    /* background: url("srcackground2.png") center; */
     background-size: cover;
 }
 
 .bg-image-createProfile {
-    background: url("srcackground3.png") center;
+    /* background: url("srcackground3.png") center; */
     background-size: cover;
 }
