+++ react/src/pages/Privacy/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Footer, SideBar, NavBar } from "../../components";
+
+const PrivacyPage = () => (
+  <>
+    <React.Fragment>
+      <h1>Privacy Policy</h1>
+      <p>Use this page to detail your site's privacy policy.</p>
+    </React.Fragment>
+  </>
+);
+
+export default PrivacyPage;
